Fixed some bugs with encoder function
Now the key should work with multiple  letters. Though caution should be taken with the new format required.

The shift function should work with negative numbers now. Which will occur if the letter of a key has a higher index than the corresponding number.

Should still check manually that the results are correct and that the functions work as intended.