Cleared some items from Notes.txt
Most of the changes should be self-explanatory with the comments that I added. The decision to make CoC like it now is was made partly for ease of coding and partly because I wouldn't expect two students to write up an official document detailing what is and what isn't allowed on the webpage.